Update the DisplayLicense serialisation tests
